맵 테마 2까지 진행

+++ Source/DragDown/Actor/DDGimmickActor.h
@@ -38,7 +38,7 @@ protected:
 	int32 NowRot;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Gimmick")
-	int32 ActionTime;
+	float ActionTime;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Gimmick")
 	bool bIsEaseOut;
